Create two_sum_problem.py
This is a function that serves to provide a solution to the classic "Two Sum Problem" in which we must return the indexes each pair of integars in a given list that sum to a given target.

Input any list of integars and any target integar and this function will return a valid solution.